Refactor game interfaces for easier maintenance
